move comms files to comms dir and update namespace
