+++ src/CMakeLists.txt
@@ -26,6 +26,11 @@ set(network_sources
   network/PeerTracker.cc
 )
 
+set(comms_sources
+  comms/Broker.cc
+  comms/MsgManager.cc
+)
+
 set(gui_sources
   ${gui_sources}
   PARENT_SCOPE
@@ -47,13 +52,13 @@ target_link_libraries(${ign_lib_target}
 set (sources
   Barrier.cc
   BaseView.cc
-  Broker.cc
+#  Broker.cc
   Conversions.cc
   EntityComponentManager.cc
   LevelManager.cc
   Link.cc
   Model.cc
-  MsgManager.cc
+#  MsgManager.cc
   Primitives.cc
   SdfEntityCreator.cc
   SdfGenerator.cc
